feat: server вынести код Run() в app
