+++ README.md
@@ -21,7 +21,7 @@
 + runtime, sync
 + zerolog
 
-Это учебная работе в [Яндекс-практикуме](https://practicum.yandex.ru), поэтому такие вещи как сжатие ответа и запроса, подписывание реализовано ручками при помощи стандартных библиотек. 
+Это учебная работе в [Яндекс-практикуме](https://practicum.yandex.ru), поэтому такие вещи как сжатие ответа и запроса, подписывание реализовано ручками при помощи стандартных библиотек.
 
 ## Вещи которыми я горжусь
